requirements.txt
python-can>=4.3